// src/code/enums.ts
/*
    * Home Control Hub
    * 
    * This file contains the enums for the Home Control Hub application.
    * Each enum is responsible for a specific aspect of the application's functionality.
    *
*/

export enum UserRole_ENUM {
    ADMIN = "ADMIN",
    STANDARD_USER = "STANDARD_USER",
    GUEST = "GUEST"
}